requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
selenium
pillow
numpy
tenacity
httpx
pytest
pytest-asyncio
pytest-mock
slowapi
sqlalchemy
alembic
python-dotenv
python-multipart
requests